ci, test/e2e: multik8s: parallelize tests on CI
