+++ test/e2e/multik8s/testutil/testdata/mantlerestore-template.yaml
