+++ test/e2e/multik8s/testutil/testdata/mount-deploy-template.yaml
