+++ test/e2e/multik8s/testutil/testdata/rbd-pool-sc-template.yaml
